requirements.txt
datasets
sentence-transformers
psycopg2-binary
python-dotenv